pyproject.toml
[tool.poetry]
name = "smartsquash"
version = "0.1.3"
description = "Makes daily git workflows easier, automates rebases or fixups."
authors = ["Max Wittig <max.wittig95@gmail.com>"]
readme = "README.md"
repository = "https://github.com/max-wittig/smartsquash"
keywords = ["git", "developers", "squash", "rebase"]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development",
    "Topic :: Utilities"
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
gitpython = "^3.0.5"
loguru = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.3.5"
coverage = "^5.0.3"

[tool.poetry.scripts]
sq = "smartsquash.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
